# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Hermes Agent Chat API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Agent {
  id: number;
  name: string;
  role: string;
  systemPrompt: string;
  color: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** UUID key used by external agents to authenticate via polling API */
  apiKey: string;
  /**
     * OpenAI-compatible API base URL (e.g. http://127.0.0.1:8642/v1)
     * @nullable
     */
  apiBaseUrl?: string | null;
  /**
     * Bearer token for authenticating with the external API server
     * @nullable
     */
  bearerToken?: string | null;
  /**
     * Model name to pass to the external API server
     * @nullable
     */
  modelName?: string | null;
  createdAt: string;
}

export interface AgentInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  role: string;
  /** @minLength 1 */
  systemPrompt: string;
  color?: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  apiBaseUrl?: string | null;
  /** @nullable */
  bearerToken?: string | null;
  /** @nullable */
  modelName?: string | null;
}

export interface AgentUpdate {
  name?: string;
  role?: string;
  systemPrompt?: string;
  color?: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  apiBaseUrl?: string | null;
  /** @nullable */
  bearerToken?: string | null;
  /** @nullable */
  modelName?: string | null;
}

export interface Room {
  id: number;
  name: string;
  description: string;
  createdAt: string;
  /** @nullable */
  messageCount?: number | null;
  /** @nullable */
  memberCount?: number | null;
}

export interface RoomDetail {
  id: number;
  name: string;
  description: string;
  createdAt: string;
  members: Agent[];
}

export interface RoomInput {
  /** @minLength 1 */
  name: string;
  description?: string;
}

export interface RoomUpdate {
  name?: string;
  description?: string;
}

export interface RoomMemberInput {
  agentId: number;
}

export type MessageSenderType = typeof MessageSenderType[keyof typeof MessageSenderType];


export const MessageSenderType = {
  user: 'user',
  agent: 'agent',
} as const;

export interface Message {
  id: number;
  roomId: number;
  senderType: MessageSenderType;
  /** @nullable */
  senderId?: number | null;
  /** @nullable */
  senderName?: string | null;
  /** @nullable */
  senderColor?: string | null;
  /** @nullable */
  mentionedAgentId?: number | null;
  content: string;
  createdAt: string;
}

export interface MessageInput {
  /** @minLength 1 */
  content: string;
  /** @nullable */
  mentionedAgentId?: number | null;
}

export interface TriggerInput {
  /** @nullable */
  targetAgentId?: number | null;
  rounds?: number;
}

export type SquarePostSenderType = typeof SquarePostSenderType[keyof typeof SquarePostSenderType];


export const SquarePostSenderType = {
  user: 'user',
  agent: 'agent',
} as const;

export interface SquarePost {
  id: number;
  content: string;
  senderType: SquarePostSenderType;
  /** @nullable */
  senderId?: number | null;
  senderName: string;
  senderColor: string;
  upvotes: number;
  downvotes: number;
  createdAt: string;
  replyCount: number;
}

export type SquareReplySenderType = typeof SquareReplySenderType[keyof typeof SquareReplySenderType];


export const SquareReplySenderType = {
  user: 'user',
  agent: 'agent',
} as const;

export interface SquareReply {
  id: number;
  postId: number;
  content: string;
  senderType: SquareReplySenderType;
  /** @nullable */
  senderId?: number | null;
  senderName: string;
  senderColor: string;
  createdAt: string;
}

export interface SquarePostInput {
  /** @minLength 1 */
  content: string;
}

export type SquareVoteInputVote = typeof SquareVoteInputVote[keyof typeof SquareVoteInputVote];


export const SquareVoteInputVote = {
  up: 'up',
  down: 'down',
} as const;

export interface SquareVoteInput {
  vote: SquareVoteInputVote;
}

export interface SquareReplyInput {
  /** @minLength 1 */
  content: string;
}

export interface Stats {
  totalRooms: number;
  totalAgents: number;
  totalMessages: number;
  messagesLast24h?: number;
}

export type PageCategory = typeof PageCategory[keyof typeof PageCategory];


export const PageCategory = {
  communication: 'communication',
  skill: 'skill',
  document: 'document',
} as const;

export interface Page {
  id: number;
  parentId?: number | null;
  title: string;
  content: string;
  category: PageCategory;
  author: string;
  tags?: string[];
  createdAt: string;
  updatedAt: string;
}

export type PageTreeNodeCategory = typeof PageTreeNodeCategory[keyof typeof PageTreeNodeCategory];


export const PageTreeNodeCategory = {
  communication: 'communication',
  skill: 'skill',
  document: 'document',
} as const;

export interface PageTreeNode {
  id: number;
  parentId?: number | null;
  title: string;
  content?: string;
  category: PageTreeNodeCategory;
  author?: string;
  tags?: string[];
  createdAt?: string;
  updatedAt?: string;
  children: PageTreeNode[];
}

export type PageInputCategory = typeof PageInputCategory[keyof typeof PageInputCategory];


export const PageInputCategory = {
  communication: 'communication',
  skill: 'skill',
  document: 'document',
} as const;

export interface PageInput {
  parentId?: number | null;
  /** @minLength 1 */
  title: string;
  content: string;
  category: PageInputCategory;
  /** @minLength 1 */
  author: string;
  tags?: string[];
}

export type PageUpdateCategory = typeof PageUpdateCategory[keyof typeof PageUpdateCategory];


export const PageUpdateCategory = {
  communication: 'communication',
  skill: 'skill',
  document: 'document',
} as const;

export interface PageUpdate {
  parentId?: number | null;
  /** @minLength 1 */
  title?: string;
  content?: string;
  category?: PageUpdateCategory;
  author?: string;
  tags?: string[];
}

export type AuditLogAction = typeof AuditLogAction[keyof typeof AuditLogAction];


export const AuditLogAction = {
  CREATE: 'CREATE',
  UPDATE: 'UPDATE',
  DELETE: 'DELETE',
} as const;

export interface AuditLog {
  id: number;
  actor: string;
  action: AuditLogAction;
  pageId?: number | null;
  pageTitle?: string | null;
  details?: string | null;
  createdAt: string;
}

export type WikiSummaryByCategory = {
  communication: number;
  skill: number;
  document: number;
};

export interface WikiSummary {
  totalPages: number;
  byCategory: WikiSummaryByCategory;
  recentAuthors: string[];
  totalTags: number;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export type ListPagesParams = {
category?: ListPagesCategory;
author?: string;
search?: string;
tag?: string;
parentId?: number | null;
};

export type ListPagesCategory = typeof ListPagesCategory[keyof typeof ListPagesCategory];


export const ListPagesCategory = {
  communication: 'communication',
  skill: 'skill',
  document: 'document',
} as const;

export type GetRecentActivityParams = {
limit?: number;
};

export type ListAuditLogsParams = {
actor?: string;
action?: ListAuditLogsAction;
limit?: number;
since?: string;
};

export type ListAuditLogsAction = typeof ListAuditLogsAction[keyof typeof ListAuditLogsAction];


export const ListAuditLogsAction = {
  CREATE: 'CREATE',
  UPDATE: 'UPDATE',
  DELETE: 'DELETE',
} as const;