# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hermes Agent Chat API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pages:
    get:
      operationId: listPages
      tags: [pages]
      summary: List all wiki pages
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [communication, skill, document]
        - name: author
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: parentId
          in: query
          schema:
            type: integer
            nullable: true
      responses:
        "200":
          description: List of wiki pages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Page"
    post:
      operationId: createPage
      tags: [pages]
      summary: Create a new wiki page
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
...
  /pages/{id}:
    get:
      operationId: getPage
      tags: [pages]
      summary: Get a wiki page by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
        "404":
          description: Not found
    patch:
      operationId: updatePage
      tags: [pages]
      summary: Update a wiki page
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PageUpdate"
      responses:
        "200":
          description: Updated page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
        "404":
          description: Not found
    delete:
      operationId: deletePage
      tags: [pages]
      summary: Delete a wiki page
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /activity/recent:
    get:
      operationId: getRecentActivity
      tags: [activity]
      summary: Get recently updated pages
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent pages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Page"
...
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'

  # ── Stats ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Page:
      type: object
      required: [id, title, content, category, author, createdAt, updatedAt]
      properties:
        id:
          type: integer
        parentId:
          type: integer
          nullable: true
        title:
          type: string
        content:
          type: string
        category:
          type: string
          enum: [communication, skill, document]
        author:
          type: string
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PageInput:
      type: object
      required: [title, content, category, author]
      properties:
        parentId:
          type: integer
          nullable: true
        title:
          type: string
          minLength: 1
        content:
          type: string
        category:
          type: string
          enum: [communication, skill, document]
        author:
          type: string
          minLength: 1
        tags:
          type: array
          items:
            type: string
    PageUpdate:
      type: object
      properties:
        parentId:
          type: integer
          nullable: true
        title:
          type: string
          minLength: 1
        content:
          type: string
        category:
          type: string
          enum: [communication, skill, document]
        author:
          type: string
        tags:
          type: array
          items:
            type: string
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
        metadata:
          $ref: '#/components/schemas/UploadUrlRequest'