# lib/api-zod/src/generated/types/pageInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Hermes Agent Chat API
 * OpenAPI spec version: 0.1.0
 */
import type { PageInputCategory } from './pageInputCategory';

export interface PageInput {
  parentId?: number | null;
  /** @minLength 1 */
  title: string;
  content: string;
  category: PageInputCategory;
  /** @minLength 1 */
  author: string;
  tags?: string[];
}