# lib/api-zod/src/generated/types/pageUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Hermes Agent Chat API
 * OpenAPI spec version: 0.1.0
 */
import type { PageUpdateCategory } from './pageUpdateCategory';

export interface PageUpdate {
  parentId?: number | null;
  /** @minLength 1 */
  title?: string;
  content?: string;
  category?: PageUpdateCategory;
  author?: string;
  tags?: string[];
}